fix: linting and path issues.
