+++ src/realtime/types/events/base.ts
@@ -1,20 +1,16 @@
 type CompanyInfo = {
-    companyId: string;
-    companyName: string;
-  };
-  
-  type Timestamps = {
-    creationDate: Date;
-    updateDate: Date;
-  };
-
-  type EventProperties = {
-    eventId: string;
-    eventTime: number;
+  companyId: string;
+  companyName: string;
 };
 
-export {
-    CompanyInfo, 
-    Timestamps, 
-    EventProperties
-}
+type Timestamps = {
+  creationDate: Date;
+  updateDate: Date;
+};
+
+type EventProperties = {
+  eventId: string;
+  eventTime: number;
+};
+
+export { CompanyInfo, Timestamps, EventProperties };
